[deploy/#126]: 렌더링을 위한 analysis 컴포넌트 분리
